Move TitleButtonLayout and ContentButtonLayout to shared component folder.
This is to be used by Setup Ads.
