+++ js/src/components/google-account/card-content.js
@@ -8,7 +8,7 @@ import { __ } from '@wordpress/i18n';
  */
 import AppSpinner from '.~/components/app-spinner';
 import useGoogleAccount from '.~/hooks/useGoogleAccount';
-import TitleButtonLayout from '../../setup-mc/setup-stepper/setup-accounts/title-button-layout';
+import TitleButtonLayout from '.~/components/title-button-layout';
 import ConnectAccount from './connect-account';
 
 const CardContent = ( props ) => {
