+++ js/src/setup-mc/setup-stepper/setup-accounts/google-mc-account/connected-card.js
@@ -2,7 +2,7 @@
  * Internal dependencies
  */
 import Section from '.~/wcdl/section';
-import TitleButtonLayout from '../title-button-layout';
+import TitleButtonLayout from '.~/components/title-button-layout';
 import AccountId from './account-id';
 
 const ConnectedCard = ( props ) => {
