+++ js/src/setup-mc/setup-stepper/setup-accounts/google-mc-account/switch-url-card/index.js
@@ -13,7 +13,7 @@ import Section from '.~/wcdl/section';
 import Subsection from '.~/wcdl/subsection';
 import useApiFetchCallback from '.~/hooks/useApiFetchCallback';
 import { useAppDispatch } from '.~/data';
-import ContentButtonLayout from '../../content-button-layout';
+import ContentButtonLayout from '.~/components/content-button-layout';
 import AccountId from '../account-id';
 import './index.scss';
