+++ js/src/setup-mc/setup-stepper/setup-accounts/wordpressdotcom-account/card-content.js
@@ -7,7 +7,7 @@ import { __ } from '@wordpress/i18n';
  * Internal dependencies
  */
 import AppSpinner from '.~/components/app-spinner';
-import TitleButtonLayout from '../title-button-layout';
+import TitleButtonLayout from '.~/components/title-button-layout';
 import ConnectAccount from './connect-account';
 import useJetpackAccount from '.~/hooks/useJetpackAccount';
